+++ src/whoosh/__init__.py
@@ -25,8 +25,8 @@
 # those of the authors and should not be interpreted as representing official
 # policies, either expressed or implied, of Matt Chaput.
 
-__version__ = (3, 0, 2)
-__version_str__ = "3.0.2"
+__version__ = (3, 0, 3)
+__version_str__ = "3.0.3"
 
 
 def versionstring(build=True, extra=True):